requirements.txt
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
jupyter>=1.0.0
huggingface_hub
datasets
Pillow
matplotlib
requests
opencv-python
